4.21*

+++ part4/bloglist/controllers/blogs.js
@@ -34,8 +34,18 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const tokenUser = request.token === undefined
+    ? false
+    : jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(400).json({ 'error': 'blog not found, malformatted id' })
+  }
+  if ( tokenUser.id && blog.user && (blog.user.toString() === tokenUser.id.toString()) ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  return response.status(401).json({ 'error': 'user unauthorized to delete the blog' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
